Revert "Add validation for single photo download"
