+++ backend/src/controlmat.Api/Controllers/PhotosController.cs
@@ -17,7 +17,10 @@ public class PhotosController : ControllerBase
     [HttpGet("{photoId}/download")]
     public async Task<IActionResult> DownloadPhoto(int photoId)
     {
-        var result = await _mediator.Send(new DownloadSinglePhotoQuery.Request(photoId));
+        var result = await _mediator.Send(new DownloadPhotoQuery.Request(photoId));
+        if (result == null)
+            return NotFound();
+
         return File(result.FileBytes, result.ContentType, result.FileName);
     }
 }
