--- backend/src/controlmat.Application/Common/Exceptions/ValidationException.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace Controlmat.Application.Common.Exceptions;
-
-public class ValidationException : Exception
-{
-    public ValidationException(string message) : base(message)
-    {
-    }
-}
