+++ backend/src/controlmat.Infrastructure/Repositories/PhotoRepository.cs
@@ -17,11 +17,6 @@ namespace Controlmat.Infrastructure.Repositories
             _context = context;
         }
 
-        public async Task<bool> ExistsAsync(int photoId)
-        {
-            return await _context.Photos.AnyAsync(p => p.Id == photoId);
-        }
-
         public async Task<Photo?> GetByIdAsync(int photoId)
         {
             return await _context.Photos.FirstOrDefaultAsync(p => p.Id == photoId);
